Move common code and tests to Configurator and ConfiguratorTest
